Move KeyboardWedge scanner implementation to the right file name
