+++ lib/alexandria/scanners/keyboard_wedge.rb
